Started multiplayer WIP

+++ src/com/dazkins/triad/game/world/tile/TileGrass.java
@@ -1,7 +1,7 @@
 package com.dazkins.triad.game.world.tile;
 
 public class TileGrass extends Tile {
-	public TileGrass(int i) {
+	public TileGrass(byte i) {
 		super(i, "Grass", 0, false, 0);
 	}
 }
