+++ src/com/dazkins/triad/game/world/tile/TileSand.java
@@ -1,7 +1,7 @@
 package com.dazkins.triad.game.world.tile;
 
 public class TileSand extends Tile {
-	public TileSand(int i) {
+	public TileSand(byte i) {
 		super(i, "Sand", 2, false, 1);
 	}
 }
