+++ src/com/dazkins/triad/game/world/tile/TileStone.java
@@ -1,7 +1,7 @@
 package com.dazkins.triad.game.world.tile;
 
 public class TileStone extends Tile {
-	public TileStone(int i) {
+	public TileStone(byte i) {
 		super(i, "Stone", 1, false, 1);
 	}
 }
