+++ src/com/dazkins/triad/game/world/tile/TileWater.java
@@ -1,7 +1,7 @@
 package com.dazkins.triad.game.world.tile;
 
 public class TileWater extends Tile {
-	public TileWater(int i) {
+	public TileWater(byte i) {
 		super(i, "Water", 4, false, 2);
 	}
 }
