+++ src/com/dazkins/triad/game/world/tile/TileWhite.java
@@ -1,7 +1,7 @@
 package com.dazkins.triad.game.world.tile;
 
 public class TileWhite extends Tile {
-	public TileWhite(int i) {
+	public TileWhite(byte i) {
 		super(i, "White", 3, false, 1);
 	}
 }
